Codigos/2025/programacion IV/Tareas/8 Envio de correos asincronos con celery y keydb/requirements.txt
Flask>=2.3
Flask-Mail>=0.9.1
Celery>=5.3
redis>=5.0
python-dotenv>=1.0
